# docs/api/planner.openapi.yaml
openapi: 3.1.0
info:
  title: Planning Platform Planner API
  version: 1.0.0
servers:
  - url: http://planner-api.planning-platform.svc.cluster.local:8080
paths:
  /health/live:
    get:
      operationId: live
      responses:
        "200":
          description: Process is alive
  /health/ready:
    get:
      operationId: ready
      responses:
        "200":
          description: Database, planner migration, and checkpoint tables are ready
        "503":
          description: A required dependency or migration is unavailable
  /metrics:
    get:
      operationId: metrics
      responses:
        "200":
          description: Prometheus metrics
  /v1/plans:
    post:
      operationId: startPlan
      summary: Start or recover a planning thread idempotently
      security:
        - InternalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartPlanRequest"
      responses:
        "200":
          description: Existing idempotent result or recovered durable result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "201":
          description: Planning thread created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "401":
          description: Internal authentication failed
        "409":
          description: Request identity conflicts or the matching claim is still in progress
        "422":
          description: Request, binding digest, date, or immutable repository snapshot is invalid
  /v1/plans/{thread_id}:
    get:
      operationId: getPlan
      security:
        - InternalToken: []
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      responses:
        "200":
          description: Durable thread state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "401":
          description: Internal authentication failed
        "404":
          description: Thread not found
        "422":
          description: Thread ID is invalid
  /v1/plans/{thread_id}/artifacts:
    get:
      operationId: getPlanArtifacts
      summary: Retrieve exact hash-verified artifact bytes for internal publication
      security:
        - InternalToken: []
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      responses:
        "200":
          description: Exact durable artifact bytes and hashes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactBundle"
        "401":
          description: Internal authentication failed
        "404":
          description: Thread not found
        "409":
          description: Artifacts are not ready
        "422":
          description: Thread ID is invalid
  /v1/plans/{thread_id}/resume:
    post:
      operationId: resumePlan
      security:
        - InternalToken: []
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumePlanRequest"
      responses:
        "200":
          description: Thread resumed, recovered, or an existing resume replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "401":
          description: Internal authentication failed
        "404":
          description: Thread not found
        "409":
          description: Comment is stale, already bound, mismatched, or currently in progress
        "422":
          description: Thread ID, request, or date is invalid
  /v1/plans/{thread_id}/abandon-terminal-resume:
    post:
      operationId: abandonTerminalResume
      summary: Abandon an expired resume owned by one terminal delivery
      security:
        - InternalToken: []
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbandonTerminalResumeRequest"
      responses:
        "200":
          description: Exact pre-resume interrupt restored and old claim tombstoned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "401":
          description: Internal authentication failed
        "404":
          description: Thread not found
        "409":
          description: Terminal resume cannot be safely abandoned
        "422":
          description: Thread ID or request is invalid
  /v1/plans/{thread_id}/convert-failed-critique-to-interrupt:
    post:
      operationId: convertFailedCritiqueToInterrupt
      summary: Convert Failed Critique To Interrupt
      security:
        - InternalToken: []
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertFailedCritiqueRequest"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "401":
          description: Internal authentication failed
        "404":
          description: Thread not found
        "409":
          description: Failed critique cannot be safely corrected
        "422":
          description: Validation Error
components:
  securitySchemes:
    InternalToken:
      type: apiKey
      in: header
      name: X-Planning-Internal-Token
      description: Internal credential for authenticated planner operations.
  parameters:
    ThreadId:
      in: path
      name: thread_id
      required: true
      schema:
        type: string
        pattern: "^openproject:[1-9][0-9]*:planning:[1-9][0-9]*$"
  schemas:
    PlannerEvent:
      type: object
      required: [idempotency_key, trace_id]
      properties:
        idempotency_key:
          type: string
          minLength: 8
          maxLength: 255
        trace_id:
          type: string
          format: uuid
      additionalProperties: true
    IdeaSnapshot:
      type: object
      additionalProperties: false
      required: [work_package_id, lock_version, updated_at, title]
      properties:
        work_package_id:
          type: integer
          minimum: 1
        lock_version:
          type: integer
          minimum: 0
        updated_at:
          type: string
          format: date-time
        title:
          type: string
          minLength: 1
        description:
          type: string
          default: ""
          maxLength: 65536
    OpenProjectSnapshotInput:
      type: object
      additionalProperties: false
      required: [captured_at, etag, sha256]
      properties:
        captured_at:
          type: string
          format: date-time
        etag:
          type: string
          minLength: 1
        sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
    RepositoryFile:
      type: object
      additionalProperties: false
      required: [path, sha256, content]
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 512
        sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        content:
          type: string
          maxLength: 262144
    RepositorySnapshot:
      type: object
      additionalProperties: false
      required: [name, commit, snapshot_sha256, files]
      properties:
        name:
          type: string
          pattern: "^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$"
        commit:
          type: string
          pattern: "^[0-9a-f]{40}$"
        snapshot_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        files:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/RepositoryFile"
    StartPlanRequest:
      type: object
      additionalProperties: false
      required:
        - event
        - idea
        - plan_id
        - plan_version
        - idea_sha256
        - openproject_snapshot
        - repositories
      properties:
        event:
          $ref: "#/components/schemas/PlannerEvent"
        idea:
          $ref: "#/components/schemas/IdeaSnapshot"
        plan_id:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,63}$"
        plan_version:
          type: integer
          minimum: 1
        idea_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        openproject_snapshot:
          $ref: "#/components/schemas/OpenProjectSnapshotInput"
        repositories:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/RepositorySnapshot"
        replan:
          anyOf:
            - $ref: "#/components/schemas/ReplanContext"
            - type: "null"
    ReplanContext:
      properties:
        prior_plan:
          $ref: "#/components/schemas/BacklogPlan"
        base_approved_planning_commit:
          type: string
          pattern: "^[0-9a-f]{40}$"
        selected_root_keys:
          items:
            type: string
            maxLength: 96
            minLength: 3
            pattern: "^[a-z][a-z0-9]*(?:-[a-z0-9]+)*$"
          type: array
          minItems: 1
          uniqueItems: true
        affected_node_keys:
          items:
            type: string
            maxLength: 96
            minLength: 3
            pattern: "^[a-z][a-z0-9]*(?:-[a-z0-9]+)*$"
          type: array
          minItems: 1
          uniqueItems: true
        reason:
          type: string
          maxLength: 4096
          minLength: 1
      additionalProperties: false
      type: object
      required:
        - prior_plan
        - base_approved_planning_commit
        - selected_root_keys
        - affected_node_keys
        - reason
      description: Immutable prior plan and the exact subtree an operator authorized to change.
    BacklogPlan:
      properties:
        schema_version:
          type: string
          const: "1.0.0"
        plan:
          $ref: "#/components/schemas/Plan"
        items:
          items:
            $ref: "#/components/schemas/BacklogItem"
          type: array
      additionalProperties: false
      type: object
      required: [schema_version, plan, items]
    Plan:
      properties:
        id:
          type: string
        version:
          type: integer
        publication_identity:
          type: string
        source_idea:
          $ref: "#/components/schemas/SourceIdea"
        repositories:
          items:
            $ref: "#/components/schemas/Repository"
          type: array
        approved_planning_commit:
          anyOf:
            - type: string
            - type: "null"
        openproject_snapshot:
          $ref: "#/components/schemas/SnapshotReference"
        replan:
          anyOf:
            - $ref: "#/components/schemas/ReplanScope"
            - type: "null"
      additionalProperties: false
      type: object
      required:
        - id
        - version
        - publication_identity
        - source_idea
        - repositories
        - approved_planning_commit
        - openproject_snapshot
    SourceIdea:
      properties:
        work_package_id:
          type: integer
        lock_version:
          type: integer
        updated_at:
          type: string
      additionalProperties: false
      type: object
      required: [work_package_id, lock_version, updated_at]
    Repository:
      properties:
        name:
          type: string
        commit:
          type: string
      additionalProperties: false
      type: object
      required: [name, commit]
    SnapshotReference:
      properties:
        captured_at:
          type: string
        etag:
          type: string
        sha256:
          type: string
      additionalProperties: false
      type: object
      required: [captured_at, etag, sha256]
    ReplanScope:
      properties:
        base_plan_version:
          type: integer
          minimum: 1
        selected_root_keys:
          items:
            type: string
            maxLength: 96
            minLength: 3
            pattern: "^[a-z][a-z0-9]*(?:-[a-z0-9]+)*$"
          type: array
          minItems: 1
          uniqueItems: true
        affected_node_keys:
          items:
            type: string
            maxLength: 96
            minLength: 3
            pattern: "^[a-z][a-z0-9]*(?:-[a-z0-9]+)*$"
          type: array
          minItems: 1
          uniqueItems: true
        retained_node_bindings:
          items:
            $ref: "#/components/schemas/ReplanNodeBinding"
          type: array
          uniqueItems: true
      additionalProperties: false
      type: object
      required:
        - base_plan_version
        - selected_root_keys
        - affected_node_keys
        - retained_node_bindings
      description: Artifact-visible boundary and provenance for a partial replan.
    ReplanNodeBinding:
      properties:
        node_key:
          type: string
          maxLength: 96
          minLength: 3
          pattern: "^[a-z][a-z0-9]*(?:-[a-z0-9]+)*$"
        plan_version:
          type: integer
          minimum: 1
        planning_commit:
          type: string
          pattern: "^[0-9a-f]{40}$"
      additionalProperties: false
      type: object
      required: [node_key, plan_version, planning_commit]
      description: Managed-field provenance retained for one protected prior node.
    BacklogItem:
      properties:
        key:
          type: string
        type:
          type: string
          enum: [Initiative, Epic, Story, Task, Decision, Investigation, Bug]
        title:
          type: string
        objective:
          type: string
        parent:
          anyOf:
            - type: string
            - type: "null"
        repository:
          type: string
        integration_work:
          type: boolean
        source_requirements:
          items:
            type: string
          type: array
        maintenance_objectives:
          items:
            type: string
          type: array
        acceptance_criteria:
          items:
            $ref: "#/components/schemas/AcceptanceCriterion"
          type: array
        blocked_by:
          items:
            type: string
          type: array
        sequence_after:
          items:
            type: string
          type: array
        related_to:
          items:
            type: string
          type: array
        decision_required:
          items:
            type: string
          type: array
        decisions:
          items:
            type: string
          type: array
        mutex:
          items:
            type: string
          type: array
        risk:
          type: string
          enum: [low, medium, high, critical]
        estimate:
          type: string
          enum: [XS, S, M]
        agent_eligibility:
          $ref: "#/components/schemas/AgentEligibility"
        validation_commands:
          items:
            type: string
          type: array
        required_evidence:
          items:
            $ref: "#/components/schemas/RequiredEvidence"
          type: array
        result_predicate:
          $ref: "#/components/schemas/ResultPredicate"
      additionalProperties: false
      type: object
      required:
        - key
        - type
        - title
        - objective
        - parent
        - repository
        - integration_work
        - source_requirements
        - maintenance_objectives
        - acceptance_criteria
        - blocked_by
        - sequence_after
        - related_to
        - decision_required
        - decisions
        - mutex
        - risk
        - estimate
        - agent_eligibility
        - validation_commands
        - required_evidence
        - result_predicate
    AcceptanceCriterion:
      properties:
        criterion:
          type: string
        observation:
          type: string
      additionalProperties: false
      type: object
      required: [criterion, observation]
    AgentEligibility:
      properties:
        eligible:
          type: boolean
        reason:
          type: string
      additionalProperties: false
      type: object
      required: [eligible, reason]
    RequiredEvidence:
      properties:
        kind:
          type: string
          enum:
            - artifact
            - check_run
            - command_output
            - decision_record
            - deployment
            - metric
            - pull_request
            - review
        description:
          type: string
      additionalProperties: false
      type: object
      required: [kind, description]
    ResultPredicate:
      properties:
        kind:
          type: string
          enum: [artifact, command, deployment, github_check, metric]
        expression:
          type: string
      additionalProperties: false
      type: object
      required: [kind, expression]
    ResumePlanRequest:
      type: object
      additionalProperties: false
      required: [event, interrupt_id, comment_id, comment_created_at, answer]
      properties:
        event:
          $ref: "#/components/schemas/PlannerEvent"
        interrupt_id:
          type: string
          minLength: 1
        comment_id:
          type: integer
          minimum: 1
        comment_created_at:
          type: string
          format: date-time
        answer:
          type: string
          minLength: 1
          maxLength: 65536
    AbandonTerminalResumeRequest:
      type: object
      additionalProperties: false
      required:
        - idempotency_key
        - interrupt_id
        - comment_id
        - comment_created_at
        - operator
        - reason
      properties:
        idempotency_key:
          type: string
          minLength: 8
          maxLength: 255
        interrupt_id:
          type: string
          minLength: 1
          maxLength: 255
        comment_id:
          type: integer
          minimum: 1
        comment_created_at:
          type: string
          format: date-time
        operator:
          type: string
          minLength: 1
          maxLength: 255
        reason:
          type: string
          minLength: 1
          maxLength: 1024
    ConvertFailedCritiqueRequest:
      type: object
      additionalProperties: false
      required:
        - idempotency_key
        - interrupt_id
        - comment_id
        - comment_created_at
        - operator
        - reason
      properties:
        idempotency_key:
          type: string
          minLength: 8
          maxLength: 255
        interrupt_id:
          type: string
          minLength: 1
          maxLength: 255
        comment_id:
          type: integer
          minimum: 1
        comment_created_at:
          type: string
          format: date-time
        operator:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9_.:@/-]{0,254}$"
        reason:
          type: string
          minLength: 1
          maxLength: 1024
    ArtifactManifestEntry:
      type: object
      additionalProperties: false
      required: [path, sha256]
      properties:
        path:
          type: string
        sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        required:
          type: boolean
          default: true
    PendingInterrupt:
      type: object
      additionalProperties: false
      required: [interrupt_id, questions, impact, created_at]
      properties:
        interrupt_id:
          type: string
        questions:
          type: array
          maxItems: 16
          items:
            type: string
            minLength: 1
            maxLength: 2048
        impact:
          type: string
          minLength: 1
          maxLength: 1024
        created_at:
          type: string
          format: date-time
    PlanResponse:
      type: object
      additionalProperties: false
      required: [thread_id, status, trace_id, interrupt, artifact_manifest]
      properties:
        thread_id:
          type: string
        status:
          type: string
          enum: [planning, needs_input, artifacts_ready, failed]
        trace_id:
          type: string
        interrupt:
          anyOf:
            - $ref: "#/components/schemas/PendingInterrupt"
            - type: "null"
        artifact_manifest:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactManifestEntry"
    ArtifactContent:
      type: object
      additionalProperties: false
      required: [path, sha256, content]
      properties:
        path:
          type: string
        sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        content:
          type: string
    ArtifactBundle:
      type: object
      additionalProperties: false
      required: [thread_id, artifacts]
      properties:
        thread_id:
          type: string
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactContent"